Split interface.artifacts into the cache, the manifest, and the storage
